Fix performance regression in bufferSliding

+++ monix-execution/shared/src/main/scala/monix/execution/internal/math.scala
@@ -56,7 +56,7 @@ private[monix] object math {
     */
   def roundToPowerOf2(nr: Long) = {
     require(nr >= 0, "nr must be positive")
-    val bit = round(log2(nr))
+    val bit = round(log2(nr.toDouble))
     1L << (if (bit > 62) 62 else bit.toInt)
   }
 
@@ -84,7 +84,7 @@ private[monix] object math {
     */
   def nextPowerOf2(nr: Long) = {
     require(nr >= 0, "nr must be positive")
-    val bit = ceil(log2(nr))
+    val bit = ceil(log2(nr.toDouble))
     1L << (if (bit > 62) 62 else bit.toInt)
   }
 }
